Repairs, Ya la transferencia se hace
Al intentar hacer la conversion de la transferencia me di cuenta que el valor en colones de la moneda era un bigdecimal y el programa no me dejaba realizar el calculo entonces lo cambie en la base de datos por double, de una vez aproveche para agregarle el interes a la moneda y el nombre a la misma por que una cosa es el codigo (CRC) y otra el nombre (Colon Costarricense) en fin ahi le pase la nueva base de datos tendrias que eliminar todos los registros y la base de datos para que no te falle, ahi agregue un script para no iniciar desde 0.
Ya la transferencia se hace lo que pasa es que no se muestra el "Transferencia realizada", ahi voy trabajando en eso XD

+++ src/java/bank/presentation/client/transfers/Model.java
@@ -15,7 +15,7 @@ public class Model {
     private List<Cuenta> destination_accounts;
     private Cuenta origin_account;
     private Cuenta destination_account;
-    private float monto;
+    private double monto;
     
     public Model() {
         origin_accounts = new ArrayList<>();
@@ -56,11 +56,11 @@ public class Model {
         this.destination_account = destination_account;
     }
 
-    public float getMonto() {
+    public double getMonto() {
         return monto;
     }
 
-    public void setMonto(float monto) {
+    public void setMonto(double monto) {
         this.monto = monto;
     }
